+++ src/components/App.tsx
@@ -1,11 +1,11 @@
 import { Outlet } from 'react-router-dom'
 import { Container } from 'semantic-ui-react'
-import AppMedia from '../components/Media'
-import Title from '../components/Title'
 import { MediaType } from '../types/Media'
 import './App.css'
 import AppFooter from './AppFooter'
 import AppHeader from './AppHeader'
+import AppMedia from './Media'
+import Title from './Title'
 
 const App = () => {
