+++ src/main.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './pages/App.tsx'
-import 'semantic-ui-css/semantic.min.css'
 import { RouterProvider, createBrowserRouter } from 'react-router-dom'
+import 'semantic-ui-css/semantic.min.css'
+import App from './components/App.tsx'
+import { ABOUT_ROUTE, CONTACT_ROUTE, GALLERY_ROUTE, HOME_ROUTE, PRICES_ROUTE } from './constants/Constants.ts'
+import About from './pages/About.tsx'
+import Contact from './pages/Contact.tsx'
 import ErrorPage from './pages/ErrorPage.tsx'
 import Gallery from './pages/Gallery.tsx'
 import Home from './pages/Home.tsx'
 import Prices from './pages/Prices.tsx'
-import Contact from './pages/Contact.tsx'
-import About from './pages/About.tsx'
-import { ABOUT_ROUTE, CONTACT_ROUTE, GALLERY_ROUTE, HOME_ROUTE, PRICES_ROUTE } from './constants/Constants.ts'
 
 const router = createBrowserRouter([
   {
